Refactor data models, add handlers for SQS events
