+++ db/cravack_activity.go
@@ -1,0 +1,6 @@
+package db
+
+type CravackActivityEvent struct {
+	*StravaEvent
+	SlackMessageTimestamp string `dynamodbav:"SlackMessageTimestamp"`
+}
